Added AddNewFuelTransaction functionality, Fuel transaction is successfully added to fueltransaction arraylist. Almost implemented viewFuelTransactions but recycler view needs a bit of work.

+++ app/src/main/java/com/example/hello_android/ServiceTransaction.java
@@ -5,10 +5,10 @@ import java.util.Date;
 
 public class ServiceTransaction extends Transaction {
 
-   private BigDecimal laborCost;
-   private BigDecimal partCost;
+   private double laborCost;
+   private double partCost;
 
-   public ServiceTransaction(Date transactionDate, BigDecimal totalCost, BigDecimal laborCost, BigDecimal partCost) {
+   public ServiceTransaction(Date transactionDate, double totalCost, double laborCost, double partCost) {
        super(transactionDate, totalCost);
        this.laborCost = laborCost;
        this.partCost = partCost;
